fix(db): refactor types module
avoid clashes with builtin types module

Fixes: #60
