+++ src/incendium/user.py
@@ -19,7 +19,7 @@ import system.user
 from com.inductiveautomation.ignition.common.user import ContactInfo, PyUser
 
 from incendium.exceptions import ApplicationError
-from incendium.types import String
+from incendium.helper.types import String
 
 
 class IncendiumUser(object):
